refactor: extract get api
