+++ frontend/src/blog/api/getApi.js
@@ -1,0 +1,8 @@
+import axios from "axios";
+
+export const getApi = ({url}) => {
+
+    return axios.get(url)
+        .then(response => response)
+        .catch(reason => console.log(reason));
+}
